Support distinct types in schema expansion. Move to DataTypeUtils.
